# docs/engineering/openapi/S1-Auth-Onboarding.yaml
openapi: 3.1.0
info:
  title: IELTS App S1 Auth & Onboarding API
  version: 0.1.0
  description: |
    Sprint S1 API contract baseline covering registration/login, session refresh/logout,
    and onboarding goal setup with assessment trigger.
servers:
  - url: http://localhost:8787
paths:
  /v1/auth/register:
    post:
      summary: Register user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              oneOf:
                - required: [email]
                - required: [phone]
              properties:
                email:
                  type: string
                  format: email
                phone:
                  type: string
                  pattern: ^\+?[0-9]{6,20}$
                password:
                  type: string
                  minLength: 8
                display_name:
                  type: string
                  minLength: 1
                  maxLength: 64
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                required: [user_id]
                properties:
                  user_id:
                    type: string
                    format: uuid
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          description: Email or phone already exists
  /v1/auth/login:
    post:
      summary: Login with email/phone and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [identifier, password]
              properties:
                identifier:
                  type: string
                password:
                  type: string
                  minLength: 1
                device_id:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/TokenResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid credentials
        '429':
          description: Too many failed attempts
  /v1/auth/refresh:
    post:
      summary: Refresh session token pair
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
                  minLength: 20
                device_id:
                  type: string
      responses:
        '200':
          $ref: '#/components/responses/TokenResponse'
        '401':
          description: Invalid refresh token
  /v1/auth/logout:
    post:
      summary: Logout current session
      security:
        - bearerAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
                  minLength: 20
      responses:
        '200':
          description: Logout success
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
        '401':
          description: Unauthorized
  /v1/users/onboarding:
    post:
      summary: Submit onboarding goal and enqueue assessment
      security:
        - bearerAuth: []
      parameters:
        - in: header
          name: Idempotency-Key
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [target_overall_band, target_exam_date, weekly_study_hours, weak_skills]
              properties:
                target_overall_band:
                  type: number
                  minimum: 0
                  maximum: 9
                  multipleOf: 0.5
                target_exam_date:
                  type: string
                  format: date
                weekly_study_hours:
                  type: integer
                  minimum: 1
                  maximum: 80
                weak_skills:
                  type: array
                  maxItems: 4
                  items:
                    type: string
                    enum: [listening, speaking, reading, writing]
      responses:
        '202':
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                required: [assessment_id, plan_id, status]
                properties:
                  assessment_id:
                    type: string
                    format: uuid
                  plan_id:
                    type: string
                    format: uuid
                  status:
                    type: string
                    enum: [processing, completed, failed]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Unauthorized
  /v1/users/onboarding/{assessment_id}/status:
    get:
      summary: Query assessment status after onboarding submit
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: assessment_id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Assessment status
          content:
            application/json:
              schema:
                type: object
                required: [assessment_id, plan_id, status, updated_at]
                properties:
                  assessment_id:
                    type: string
                  plan_id:
                    type: string
                  status:
                    type: string
                    enum: [processing, completed, failed]
                  updated_at:
                    type: string
                    format: date-time
                  error_message:
                    type: string
        '401':
          description: Unauthorized
        '404':
          description: Assessment not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
  responses:
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            type: object
            required: [code, message]
            properties:
              code:
                type: string
                example: VALIDATION_ERROR
              message:
                type: string
    TokenResponse:
      description: Access and refresh tokens
      content:
        application/json:
          schema:
            type: object
            required: [access_token, refresh_token, expires_in, user_id, session_id]
            properties:
              access_token:
                type: string
              refresh_token:
                type: string
              expires_in:
                type: integer
              user_id:
                type: string
              session_id:
                type: string